FizzBuzz Game
Contains two methods: FizzBuzz() and FizzBuzz2()

FizzBuzz() prints through numbers 1 to 100. If one of those numbers if a multiple of 3, it prints 'Fizz'. If it is a multiple of 5, it prints 'Buzz'. If it a multiple of both, it prints 'FizzBuzz'.
FizzBuzz2() is a more versatile and complicated version of the game that allows the user to select where the numbers start, where the numbers end, what numbers are tested to see whether or not they are factors, and what is said in place of the number if it is a factor.

Currently, FizzBuzz2() tests numbers from 1 to 100. If one of the numbers is a multiple of 2, it says 'Fozz', if it is a factor of 3, it says 'Fizz', and if it is a factor of 5, it says 'Buzz'.